+++ R/imfweo-package.R
@@ -4,5 +4,5 @@
 #' @importFrom tidyr pivot_longer
 #' @importFrom cli cli_alert_info cli_alert_success
 #' @importFrom readr read_tsv
-#' @importFrom rlang .data
+#' @importFrom rlang .data .env
 "_PACKAGE"