+++ R/zzz.R
@@ -1,7 +1,6 @@
 .weo_cache <- new.env(parent = emptyenv())
 
 .onLoad <- function(libname, pkgname) {
-  .weo_cache$latest_publication <- NULL
   .weo_cache$year <- NULL
   .weo_cache$release <- NULL
   .weo_cache$bulk <- NULL
@@ -14,7 +13,6 @@
 #'
 #' @export
 weo_cache_reset <- function() {
-  .weo_cache$latest_publication <- NULL
   .weo_cache$year <- NULL
   .weo_cache$release <- NULL
   .weo_cache$bulk <- NULL